requirements.txt
numpy==1.16.2
scipy==1.2.1
tensorflow-gpu==1.12.0
h5py==2.9.0
Pillow==6.0.0
